moving middleware utils in their own folder
