+++ locksmith/package.json
@@ -10,9 +10,9 @@
     "dev": "NODE_ENV=development yarn db:migrate && ts-node-dev -r ./src/utils/envLoader.js server.js",
     "test": "yarn test:db:setup && yarn test:run",
     "test:db:setup": "yarn run test:db:drop && yarn run test:db:create && yarn run test:db:migrate",
-    "test:db:create": "",
+    "test:db:create": "NODE_ENV=test yarn sequelize db:create",
     "test:db:migrate": "NODE_ENV=test yarn sequelize db:migrate",
-    "test:db:drop": "",
+    "test:db:drop": "NODE_ENV=test yarn sequelize db:drop",
     "test:run": "NODE_ENV=test jest --forceExit --runInBand",
     "lint": "eslint --resolve-plugins-relative-to ../packages/eslint-config --ext .ts,.js src/ __tests__/",
     "db:migrate": "yarn sequelize db:migrate",
